# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "رسالة الخيرية API",
    "description": "## نظام إدارة رسالة الخيرية\n\n### الأدوار المتاحة:\n| الدور | طريقة الدخول | الصلاحيات |\n|-------|-------------|----------|\n| **Donor** | Phone Number | تسجيل — عرض بياناته — الاشتراك |\n| **Reception** | Username | إدارة المتبرعين — تسجيل تبرعات |\n| **Admin** | Username | كل الصلاحيات + إدارة الموظفين |\n\n### كيفية الاستخدام:\n1. سجّل دخول من `/api/v1/auth/login`\n2. انسخ الـ `token` من الـ Response\n3. اضغط **Authorize** في الأعلى والصق: `Bearer {token}`",
    "contact": {
      "name": "Resala Dev Team",
      "email": "dev@resala.org"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/sponsorships": {
      "post": {
        "tags": [
          "Sponsorships — Admin"
        ],
        "summary": "[Admin] إنشاء برنامج كفالة جديد",
        "description": "يُنشئ برنامج كفالة جديد في النظام.\r\n            \r\n**⚠️ يتطلب Bearer Token بدور Admin**\r\n            \r\n**Request Example:**\r\n```json\r\n{\r\n  \"name\": \"كفالة يتيم\",\r\n  \"description\": \"برنامج لكفالة الأيتام وتوفير احتياجاتهم\",\r\n  \"imageUrl\": \"https://example.com/image.jpg\",\r\n  \"icon\": \"orphan-icon\",\r\n  \"targetAmount\": 50000\r\n}\r\n```\r\n            \r\n**Success Response (201):**\r\n```json\r\n{\r\n  \"id\": 1,\r\n  \"name\": \"كفالة يتيم\",\r\n  \"description\": \"برنامج لكفالة الأيتام\",\r\n  \"targetAmount\": 50000,\r\n  \"collectedAmount\": 0,\r\n  \"isActive\": true,\r\n  \"createdAt\": \"2026-03-21T00:00:00\"\r\n}\r\n```",
        "operationId": "Sponsorships_Create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "Icon": {
                    "type": "string"
                  },
                  "TargetAmount": {
                    "type": "number",
                    "format": "double"
                  },
                  "ImageFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Icon": {
                  "style": "form"
                },
                "TargetAmount": {
                  "style": "form"
                },
                "ImageFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء برنامج الكفالة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SponsorshipViewModel"
                }
              }
            }
          },
          "400": {
            "description": "بيانات غير صحيحة",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "لم يتم تقديم Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "الدور غير مصرّح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sponsorships — Public"
        ],
        "summary": "جلب كل برامج الكفالة",
        "description": "يرجع قائمة بكل برامج الكفالة المتاحة في النظام.\r\n            \r\n**متاح للجميع بدون Token**\r\n            \r\n**Success Response (200):**\r\n```json\r\n[\r\n  {\r\n    \"id\": 1,\r\n    \"name\": \"كفالة يتيم\",\r\n    \"description\": \"برنامج لكفالة الأيتام\",\r\n    \"targetAmount\": 50000,\r\n    \"collectedAmount\": 12000,\r\n    \"isActive\": true,\r\n    \"createdAt\": \"2026-03-21T00:00:00\"\r\n  }\r\n]\r\n```",
        "operationId": "Sponsorships_GetAll",
        "responses": {
          "200": {
            "description": "قائمة برامج الكفالة",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SponsorshipViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sponsorships/{id}": {
      "get": {
        "tags": [
          "Sponsorships — Public"
        ],
        "summary": "جلب برنامج كفالة بالـ ID",
        "description": "يرجع تفاصيل برنامج كفالة محدد.\r\n            \r\n**متاح للجميع بدون Token**\r\n            \r\n**Success Response (200):**\r\n```json\r\n{\r\n  \"id\": 1,\r\n  \"name\": \"كفالة يتيم\",\r\n  \"description\": \"برنامج لكفالة الأيتام\",\r\n  \"imageUrl\": \"https://example.com/image.jpg\",\r\n  \"icon\": \"orphan-icon\",\r\n  \"targetAmount\": 50000,\r\n  \"collectedAmount\": 12000,\r\n  \"isActive\": true,\r\n  \"createdAt\": \"2026-03-21T00:00:00\"\r\n}\r\n```",
        "operationId": "Sponsorships_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "رقم برنامج الكفالة",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تفاصيل برنامج الكفالة",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SponsorshipViewModel"
                }
              }
            }
          },
          "404": {
            "description": "برنامج الكفالة غير موجود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sponsorships — Admin"
        ],
        "summary": "[Admin] تعديل برنامج كفالة",
        "description": "يُعدّل بيانات برنامج كفالة موجود.\r\n            \r\n**⚠️ يتطلب Bearer Token بدور Admin**\r\n            \r\n**Request Example:**\r\n```json\r\n{\r\n  \"name\": \"كفالة يتيم — محدّث\",\r\n  \"description\": \"وصف محدّث للبرنامج\",\r\n  \"imageUrl\": \"https://example.com/new-image.jpg\",\r\n  \"icon\": \"new-icon\",\r\n  \"targetAmount\": 75000,\r\n  \"isActive\": true\r\n}\r\n```",
        "operationId": "Sponsorships_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "رقم برنامج الكفالة",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "Icon": {
                    "type": "string"
                  },
                  "TargetAmount": {
                    "type": "number",
                    "format": "double"
                  },
                  "IsActive": {
                    "type": "boolean"
                  },
                  "ImageFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Icon": {
                  "style": "form"
                },
                "TargetAmount": {
                  "style": "form"
                },
                "IsActive": {
                  "style": "form"
                },
                "ImageFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم التعديل بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SponsorshipViewModel"
                }
              }
            }
          },
          "400": {
            "description": "بيانات غير صحيحة",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "برنامج الكفالة غير موجود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "لم يتم تقديم Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "الدور غير مصرّح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sponsorships — Admin"
        ],
        "summary": "[Admin] حذف برنامج كفالة",
        "description": "يحذف برنامج كفالة من النظام.\r\n            \r\n**⚠️ يتطلب Bearer Token بدور Admin**\r\n            \r\n**ملاحظة:** الحذف نهائي — لا يمكن التراجع.",
        "operationId": "Sponsorships_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "رقم برنامج الكفالة المراد حذفه",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "تم الحذف بنجاح"
          },
          "404": {
            "description": "برنامج الكفالة غير موجود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "لم يتم تقديم Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "الدور غير مصرّح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SponsorshipViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "imagePublicId": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "targetAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "collectedAmount": {
            "type": "number",
            "format": "double"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "أدخل الـ Token هكذا: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}